Add holy grail layout
Set the flex-direction of the body to column to set the footer to stick at the bottom end of the viewport.
Sidebar stretches to the bottom
Added the cards in 1 div container to align them properly and shrink on smaller device screens.
Footer stays at the bottom and its content is horizontally and vertically centered.